MAINT: Rename `__version__.py` to `_version.py`
